Fix po subdirectory traversal in Makefiles (again)
The solution in 0e1c1901993c ("Fix po subdirectory traversal in
Makefiles") doesn't do what I thought it would do. The duplicate rules
are not merged and only the last rule is used for each target. This, in
effect, means that the translations are not built at all for gui,
python, and sandbox.

In case of "gui" and "sandbox", fix it by providing a separate set of
po-* targets and make each of "all", "install", and "clean" depend on
them.

In case of "python", split the current universal rule into two and
traverse into "po" only for "all", "install", and "clean".

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>